Keeping category rates in double-precision until conversion absolutely necessary

+++ trunk/libhmsbeagle/GPU/BeagleGPUImpl.h
@@ -85,7 +85,7 @@ private:
     GPUPtr dIntegrationTmp;
     GPUPtr dPartialsTmp;
     
-	REAL* hCategoryRates;
+	double* hCategoryRates; // Can keep in double-precision
     
     GPUPtr* dPartials;
     GPUPtr* dMatrices;
